colpali-vision-rag/backend/requirements.txt
# ColPali + Byaldi
colpali-engine>=0.3.5
byaldi>=0.1.0

# PyTorch (CPU for compatibility)
torch>=2.2.0,<2.6.0
torchvision>=0.17.0

# Transformers
transformers>=4.46.2

# Image processing
Pillow>=10.0.0
pdf2image>=1.16.0

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# OpenAI
openai>=1.12.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
